+++ R/decorate_storr.R
@@ -69,8 +69,46 @@ refclass_decorated_storr <- methods::setRefClass(
     set = function(key, value, ...) {
       dcst_set(value = value, key = key, ..., .self = .self)
     },
-    reset_ht_hash = function() {
+    memo_hash = function(x, fun, ...) {
+      ht_memo(ht = .self$ht_hash, x = x, fun = fun, ...)
+    },
+    reset_memo_hash = function() {
       ht_clear(.self$ht_hash)
+    },
+    reset_ht_hash = function() {
+      # deprecated on 2019-09-13
+      ht_clear(.self$ht_hash)
+    },
+    encode_path = function(x) {
+      ht_memo(ht = .self$ht_encode_path, x = x, fun = reencode_path)
+    },
+    decode_path = function(x) {
+      ht_memo(ht = .self$ht_decode_path, x = x, fun = redecode_path)
+    },
+    encode_namespaced = function(x) {
+      ht_memo(ht = .self$ht_encode_namespaced, x = x, fun = reencode_namespaced)
+    },
+    decode_namespaced = function(x) {
+      ht_memo(ht = .self$ht_decode_namespaced, x = x, fun = redecode_namespaced)
+    },
+    display_keys = function(x) {
+      vapply(
+        X = x,
+        FUN = display_key,
+        FUN.VALUE = character(1),
+        USE.NAMES = FALSE,
+        .self = .self
+      )
+    },
+    set_progress = function(target, value) {
+      .self$driver$set_hash(
+        key = target,
+        namespace = "progress",
+        hash = .self$ht_progress[[value]]
+      )
+    },
+    get_progress = function(target) {
+      retrieve_progress(target = target, cache = .self)
     },
     # Delegate to storr:
     archive_export = function(...) .self$storr$archive_export(...),
@@ -256,4 +294,155 @@ dir_create <- function(x) {
     stop("cannot create directory at ", shQuote(x), call. = FALSE)
   }
   invisible()
+}
+
+#' @title Show a file's encoded representation in the cache
+#' \lifecycle{stable}
+#' @description This function simply wraps literal double quotes around
+#' the argument `x` so `drake` knows it is the name of a file.
+#' Use when you are calling functions like `deps_code()`: for example,
+#' `deps_code(file_store("report.md"))`. See the examples for details.
+#' Internally, `drake` wraps the names of file targets/imports
+#' inside literal double quotes to avoid confusion between
+#' files and generic R objects.
+#' @export
+#' @return A single-quoted character string: i.e., a filename
+#' understandable by drake.
+#' @param x Character string to be turned into a filename
+#' understandable by drake (i.e., a string with literal
+#' single quotes on both ends).
+#' @examples
+#' # Wraps the string in single quotes.
+#' file_store("my_file.rds") # "'my_file.rds'"
+#' \dontrun{
+#' isolate_example("contain side effects", {
+#' if (suppressWarnings(require("knitr"))) {
+#' load_mtcars_example() # Get the code with drake_example("mtcars").
+#' make(my_plan) # Run the workflow to build the targets
+#' list.files() # Should include input "report.Rmd" and output "report.md".
+#' head(readd(small)) # You can use symbols for ordinary objects.
+#' # But if you want to read cached info on files, use `file_store()`.
+#' readd(file_store("report.md"), character_only = TRUE) # File fingerprint.
+#' deps_code(file_store("report.Rmd"))
+#' config <- drake_config(my_plan)
+#' deps_profile(
+#'   file_store("report.Rmd"),
+#'   config = config,
+#'   character_only = TRUE
+#' )
+#' }
+#' })
+#' }
+file_store <- function(x) {
+  reencode_path(x)
+}
+
+display_key <- function(x, .self) {
+  if (is_encoded_path(x)) {
+    display_path(x = x, .self = .self)
+  } else if (is_encoded_namespaced(x)) {
+    sprintf("%s", .self$decode_namespaced(x = x))
+  } else {
+    x
+  }
+}
+
+display_path <- function(x, .self) {
+  path_ <- .self$decode_path(x = x)
+  if (is_url(path_)) {
+    sprintf("url %s", path_)
+  } else {
+    sprintf("file %s", path_)
+  }
+}
+
+redisplay_keys <- function(x) {
+  vapply(
+    X = x,
+    FUN = redisplay_key,
+    FUN.VALUE = character(1),
+    USE.NAMES = FALSE
+  )
+}
+
+redisplay_key <- function(x) {
+  if (is_encoded_path(x)) {
+    redisplay_path(x = x)
+  } else if (is_encoded_namespaced(x)) {
+    sprintf("%s", redecode_namespaced(x = x))
+  } else {
+    x
+  }
+}
+
+redisplay_path <- function(x) {
+  path <- redecode_path(x = x)
+  if (is_url(path)) {
+    sprintf("url %s", path)
+  } else {
+    sprintf("file %s", path)
+  }
+}
+
+is_encoded_path <- function(x) {
+  substr(x = x, start = 1, stop = 2) == "p-"
+}
+
+is_encoded_namespaced <- function(x) {
+  substr(x = x, start = 1, stop = 2) == "n-"
+}
+
+reencode_path <- function(x) {
+  y <- base64url::base32_encode(x = x, use.padding = FALSE)
+  sprintf("p-%s", y)
+}
+
+redecode_path <- function(x) {
+  y <- substr(x = x, start = 3, stop = 1e6)
+  base64url::base32_decode(x = y, use.padding = FALSE)
+}
+
+reencode_namespaced <- function(x) {
+  y <- base64url::base32_encode(x, use.padding = FALSE)
+  sprintf("n-%s", y)
+}
+
+redecode_namespaced <- redecode_path
+
+standardize_key <- function(text) {
+  if (any(grepl("::", text))) {
+    text <- reencode_namespaced(text)
+  }
+  text
+}
+
+ht_progress <- function(hash_algorithm) {
+  keys <- c("running", "done", "failed")
+  out <- lapply(keys, progress_hash, hash_algorithm = hash_algorithm)
+  names(out) <- keys
+  out
+}
+
+progress_hash <- function(key, hash_algorithm) {
+  out <- digest::digest(
+    key,
+    algo = hash_algorithm,
+    serialize = FALSE
+  )
+  gsub("^.", substr(key, 1, 1), out)
+}
+
+retrieve_progress <- function(target, cache) {
+  if (cache$exists(key = target, namespace = "progress")) {
+    hash <- cache$get_hash(key = target, namespace = "progress")
+    switch(
+      substr(hash, 1, 1),
+      r = "running",
+      d = "done",
+      f = "failed",
+      NA_character_
+    )
+  } else{
+    "none"
+  }
 }